all: rename package
Closes #68
