+++ README.md
@@ -1,12 +1,12 @@
-# Laravel Chunk Uploader
+# Laravel Upload Handler
 
-Chunk Uploader Package For Laravel
+Upload Handler Package For Laravel
 
-[![Github Actions Badge](https://github.com/coding-socks/laravel-chunk-uploader/workflows/test/badge.svg)](https://github.com/coding-socks/laravel-chunk-uploader/actions?query=workflow%3A"test")
-[![Downloads Badge](https://poser.pugx.org/coding-socks/laravel-chunk-uploader/downloads)](https://packagist.org/packages/coding-socks/laravel-chunk-uploader)
-[![Version Badge](https://poser.pugx.org/coding-socks/laravel-chunk-uploader/version)](https://packagist.org/packages/coding-socks/laravel-chunk-uploader)
-[![Coverage Badge](https://scrutinizer-ci.com/g/coding-socks/laravel-chunk-uploader/badges/coverage.png?b=master)](https://scrutinizer-ci.com/g/coding-socks/laravel-chunk-uploader/)
-[![License Badge](https://poser.pugx.org/coding-socks/laravel-chunk-uploader/license)](https://packagist.org/packages/coding-socks/laravel-chunk-uploader)
+[![Github Actions Badge](https://github.com/coding-socks/laravel-upload-handler/workflows/test/badge.svg)](https://github.com/coding-socks/laravel-upload-handler/actions?query=workflow%3A"test")
+[![Downloads Badge](https://poser.pugx.org/coding-socks/laravel-upload-handler/downloads)](https://packagist.org/packages/coding-socks/laravel-upload-handler)
+[![Version Badge](https://poser.pugx.org/coding-socks/laravel-upload-handler/version)](https://packagist.org/packages/coding-socks/laravel-upload-handler)
+[![Coverage Badge](https://scrutinizer-ci.com/g/coding-socks/laravel-upload-handler/badges/coverage.png?b=master)](https://scrutinizer-ci.com/g/coding-socks/laravel-upload-handler/)
+[![License Badge](https://poser.pugx.org/coding-socks/laravel-upload-handler/license)](https://packagist.org/packages/coding-socks/laravel-upload-handler)
 
 This package helps integrate a Laravel application with chunk uploader libraries eg.
 [DropzoneJS](https://www.dropzonejs.com/) and
@@ -53,7 +53,7 @@ project at the moment is [tus](https://tus.io/).
 You can easily install this package using Composer, by running the following command:
 
 ```bash
-composer require coding-socks/laravel-chunk-uploader
+composer require coding-socks/laravel-upload-handler
 ```
 
 ### Requirements
@@ -73,7 +73,7 @@ Route::any('/my-route', 'MyController@myFunction');
   - Using dependency injection
 ```php
 use Illuminate\Http\Request;
-use CodingSocks\ChunkUploader\UploadHandler;
+use CodingSocks\UploadHandler\UploadHandler;
 
 class MyController extends Controller
 {
@@ -86,7 +86,7 @@ class MyController extends Controller
   - Resolving from the app container
 ```php
 use Illuminate\Http\Request;
-use CodingSocks\ChunkUploader\UploadHandler;
+use CodingSocks\UploadHandler\UploadHandler;
 
 class MyController extends Controller
 {
@@ -110,7 +110,7 @@ library can be differ very much. Also, when possible the library gives the oppor
 
 ### Events
 
-Once a file upload is finished a `\CodingSocks\ChunkUploader\Event\FileUploaded` is triggered. This event contains
+Once a file upload is finished a `\CodingSocks\UploadHandler\Event\FileUploaded` is triggered. This event contains
 the disk and the path of the uploaded file.
 [Registering Events & Listeners from Laravel](https://laravel.com/docs/5.8/events#registering-events-and-listeners)
 
@@ -139,8 +139,8 @@ app()->make(UploadManager::class)->extend('my_driver', function () {
 });
 ```
 
-If you are adding a driver you need to extend the `\CodingSocks\ChunkUploader\Driver\UploadDriver` abstract class, for
-which you can use the shipped drivers (e.g. `\CodingSocks\ChunkUploader\Driver\BlueimpUploadDriver`) as an example as to
+If you are adding a driver you need to extend the `\CodingSocks\UploadHandler\Driver\UploadDriver` abstract class, for
+which you can use the shipped drivers (e.g. `\CodingSocks\UploadHandler\Driver\BlueimpUploadDriver`) as an example as to
 how.
 
 If you wrote a custom driver that others might find useful, please consider adding it to the package via a pull request.
@@ -182,7 +182,7 @@ This driver handles requests made by the DropzoneJS client library.
 
 This driver handles requests made by the Flow.js client library.
 
-Because of [Issue #44](https://github.com/coding-socks/laravel-chunk-uploader/issues/44) you must use `forceChunkSize`
+Because of [Issue #44](https://github.com/coding-socks/laravel-upload-handler/issues/44) you must use `forceChunkSize`
 option.
 
 ### ng-file-upload driver
@@ -203,7 +203,7 @@ This driver handles requests made by the Plupload client library.
 
 This driver handles requests made by the Resumable.js client library.
 
-Because of [Issue #44](https://github.com/coding-socks/laravel-chunk-uploader/issues/44) you must use `forceChunkSize`
+Because of [Issue #44](https://github.com/coding-socks/laravel-upload-handler/issues/44) you must use `forceChunkSize`
 option.
 
 ### simple-uploader.js driver
@@ -212,7 +212,7 @@ option.
 
 This driver handles requests made by the simple-uploader.js client library.
 
-Because of [Issue #44](https://github.com/coding-socks/laravel-chunk-uploader/issues/44) you must use `forceChunkSize`
+Because of [Issue #44](https://github.com/coding-socks/laravel-upload-handler/issues/44) you must use `forceChunkSize`
 option.
 
 ## Identifiers
