+++ src/Driver/DropzoneBaseHandler.php
@@ -1,90 +1,62 @@
 <?php
 
-namespace CodingSocks\ChunkUploader\Driver;
+namespace CodingSocks\UploadHandler\Driver;
 
 use Closure;
-use CodingSocks\ChunkUploader\Helper\ChunkHelpers;
-use CodingSocks\ChunkUploader\Identifier\Identifier;
-use CodingSocks\ChunkUploader\Range\NgFileUploadRange;
-use CodingSocks\ChunkUploader\Response\PercentageJsonResponse;
-use CodingSocks\ChunkUploader\StorageConfig;
-use Illuminate\Http\JsonResponse;
+use CodingSocks\UploadHandler\Helper\ChunkHelpers;
+use CodingSocks\UploadHandler\Range\DropzoneRange;
+use CodingSocks\UploadHandler\Response\PercentageJsonResponse;
+use CodingSocks\UploadHandler\StorageConfig;
 use Illuminate\Http\Request;
 use Illuminate\Http\UploadedFile;
-use Illuminate\Support\Arr;
 use InvalidArgumentException;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\HttpKernel\Exception\MethodNotAllowedHttpException;
 
-class NgFileUploadDriver extends UploadDriver
+class DropzoneBaseHandler extends BaseHandler
 {
     use ChunkHelpers;
 
     /**
-     * @var \CodingSocks\ChunkUploader\Identifier\Identifier
+     * @var string
      */
-    private $identifier;
+    private $fileParam;
 
     /**
-     * NgFileUploadDriver constructor.
+     * DropzoneDriver constructor.
      *
-     * @param \CodingSocks\ChunkUploader\Identifier\Identifier $identifier
+     * @param array $config
      */
-    public function __construct(Identifier $identifier)
+    public function __construct($config)
     {
-        $this->identifier = $identifier;
+        $this->fileParam = $config['param'];
     }
 
     /**
-     * @inheritDoc
+     * {@inheritDoc}
      */
     public function handle(Request $request, StorageConfig $config, Closure $fileUploaded = null): Response
     {
-        if ($this->isRequestMethodIn($request, [Request::METHOD_GET])) {
-            return $this->resume($request, $config);
-        }
-
         if ($this->isRequestMethodIn($request, [Request::METHOD_POST])) {
             return $this->save($request, $config, $fileUploaded);
         }
 
         throw new MethodNotAllowedHttpException([
-            Request::METHOD_GET,
             Request::METHOD_POST,
         ]);
     }
 
-    private function resume(Request $request, StorageConfig $config): Response
+    /**
+     * @param \Illuminate\Http\Request $request
+     * @param \CodingSocks\UploadHandler\StorageConfig $config
+     * @param \Closure|null $fileUploaded
+     *
+     * @return \Symfony\Component\HttpFoundation\Response
+     */
+    public function save(Request $request, StorageConfig $config, Closure $fileUploaded = null): Response
     {
-        $request->validate([
-            'file' => 'required',
-            'totalSize' => 'required',
-        ]);
-
-        $originalFilename = $request->get('file');
-        $totalSize = $request->get('totalSize');
-        $uid = $this->identifier->generateFileIdentifier($totalSize, $originalFilename);
-
-        if (!$this->chunkExists($config, $uid)) {
-            return new JsonResponse([
-                'file' => $originalFilename,
-                'size' => 0,
-            ]);
-        }
-
-        $chunk = Arr::last($this->chunks($config, $uid));
-        $size = explode('-', basename($chunk))[1] + 1;
-
-        return new JsonResponse([
-            'file' => $originalFilename,
-            'size' => $size,
-        ]);
-    }
-
-    private function save(Request $request, StorageConfig $config, Closure $fileUploaded = null): Response
-    {
-        $file = $request->file('file');
+        $file = $request->file($this->fileParam);
 
         $this->validateUploadedFile($file);
 
@@ -97,9 +69,19 @@ class NgFileUploadDriver extends UploadDriver
         return $this->saveChunk($file, $request, $config, $fileUploaded);
     }
 
-    private function isMonolithRequest(Request $request)
+    /**
+     * @param \Illuminate\Http\Request $request
+     *
+     * @return bool
+     */
+    private function isMonolithRequest(Request $request): bool
     {
-        return empty($request->post());
+        return $request->post('dzuuid') === null
+            && $request->post('dzchunkindex') === null
+            && $request->post('dztotalfilesize') === null
+            && $request->post('dzchunksize') === null
+            && $request->post('dztotalchunkcount') === null
+            && $request->post('dzchunkbyteoffset') === null;
     }
 
     /**
@@ -108,16 +90,18 @@ class NgFileUploadDriver extends UploadDriver
     private function validateChunkRequest(Request $request): void
     {
         $request->validate([
-            '_chunkNumber' => 'required|numeric',
-            '_chunkSize' => 'required|numeric',
-            '_totalSize' => 'required|numeric',
-            '_currentChunkSize' => 'required|numeric',
+            'dzuuid' => 'required',
+            'dzchunkindex' => 'required',
+            'dztotalfilesize' => 'required',
+            'dzchunksize' => 'required',
+            'dztotalchunkcount' => 'required',
+            'dzchunkbyteoffset' => 'required',
         ]);
     }
 
     /**
      * @param \Illuminate\Http\UploadedFile $file
-     * @param \CodingSocks\ChunkUploader\StorageConfig $config
+     * @param \CodingSocks\UploadHandler\StorageConfig $config
      * @param \Closure|null $fileUploaded
      *
      * @return \Symfony\Component\HttpFoundation\Response
@@ -136,7 +120,7 @@ class NgFileUploadDriver extends UploadDriver
     /**
      * @param \Illuminate\Http\UploadedFile $file
      * @param \Illuminate\Http\Request $request
-     * @param \CodingSocks\ChunkUploader\StorageConfig $config
+     * @param \CodingSocks\UploadHandler\StorageConfig $config
      * @param \Closure|null $fileUploaded
      *
      * @return \Symfony\Component\HttpFoundation\Response
@@ -144,19 +128,23 @@ class NgFileUploadDriver extends UploadDriver
     private function saveChunk(UploadedFile $file, Request $request, StorageConfig $config, Closure $fileUploaded = null): Response
     {
         try {
-            $range = new NgFileUploadRange($request);
+            $range = new DropzoneRange(
+                $request,
+                'dzchunkindex',
+                'dztotalchunkcount',
+                'dzchunksize',
+                'dztotalfilesize'
+            );
         } catch (InvalidArgumentException $e) {
             throw new BadRequestHttpException($e->getMessage(), $e);
         }
 
-        $originalFilename = $file->getClientOriginalName();
-        $totalSize = $request->get('_totalSize');
-        $uid = $this->identifier->generateFileIdentifier($totalSize, $originalFilename);
+        $uid = $request->post('dzuuid');
 
         $chunks = $this->storeChunk($config, $range, $file, $uid);
 
-        if (!$range->isLast()) {
-            return new PercentageJsonResponse($range->getPercentage());
+        if (!$range->isFinished($chunks)) {
+            return new PercentageJsonResponse($range->getPercentage($chunks));
         }
 
         $targetFilename = $file->hashName();
