+++ src/Driver/ResumableJsBaseHandler.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace CodingSocks\ChunkUploader\Driver;
+namespace CodingSocks\UploadHandler\Driver;
 
 use Closure;
-use CodingSocks\ChunkUploader\Helper\ChunkHelpers;
-use CodingSocks\ChunkUploader\Identifier\Identifier;
-use CodingSocks\ChunkUploader\Range\ResumableJsRange;
-use CodingSocks\ChunkUploader\Response\PercentageJsonResponse;
-use CodingSocks\ChunkUploader\StorageConfig;
+use CodingSocks\UploadHandler\Helper\ChunkHelpers;
+use CodingSocks\UploadHandler\Identifier\Identifier;
+use CodingSocks\UploadHandler\Range\ResumableJsRange;
+use CodingSocks\UploadHandler\Response\PercentageJsonResponse;
+use CodingSocks\UploadHandler\StorageConfig;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\UploadedFile;
@@ -16,7 +16,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\HttpKernel\Exception\MethodNotAllowedHttpException;
 
-class ResumableJsUploadDriver extends UploadDriver
+class ResumableJsBaseHandler extends BaseHandler
 {
     use ChunkHelpers;
 
@@ -26,7 +26,7 @@ class ResumableJsUploadDriver extends UploadDriver
     private $fileParam;
 
     /**
-     * @var \CodingSocks\ChunkUploader\Identifier\Identifier
+     * @var \CodingSocks\UploadHandler\Identifier\Identifier
      */
     private $identifier;
 
@@ -51,10 +51,10 @@ class ResumableJsUploadDriver extends UploadDriver
     private $parameterNames;
 
     /**
-     * ResumableJsUploadDriver constructor.
+     * ResumableJsDriver constructor.
      *
      * @param array $config
-     * @param \CodingSocks\ChunkUploader\Identifier\Identifier $identifier
+     * @param \CodingSocks\UploadHandler\Identifier\Identifier $identifier
      */
     public function __construct($config, Identifier $identifier)
     {
@@ -89,7 +89,7 @@ class ResumableJsUploadDriver extends UploadDriver
 
     /**
      * @param \Illuminate\Http\Request $request
-     * @param \CodingSocks\ChunkUploader\StorageConfig $config
+     * @param \CodingSocks\UploadHandler\StorageConfig $config
      *
      * @return \Symfony\Component\HttpFoundation\Response
      */
@@ -121,7 +121,7 @@ class ResumableJsUploadDriver extends UploadDriver
 
     /**
      * @param \Illuminate\Http\Request $request
-     * @param \CodingSocks\ChunkUploader\StorageConfig $config
+     * @param \CodingSocks\UploadHandler\StorageConfig $config
      * @param \Closure|null $fileUploaded
      *
      * @return \Symfony\Component\HttpFoundation\Response
@@ -154,7 +154,7 @@ class ResumableJsUploadDriver extends UploadDriver
     /**
      * @param \Illuminate\Http\UploadedFile $file
      * @param \Illuminate\Http\Request $request
-     * @param \CodingSocks\ChunkUploader\StorageConfig $config
+     * @param \CodingSocks\UploadHandler\StorageConfig $config
      * @param \Closure|null $fileUploaded
      *
      * @return \Symfony\Component\HttpFoundation\Response
