+++ src/Range/Range.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace CodingSocks\ChunkUploader\Range;
+namespace CodingSocks\UploadHandler\Range;
 
 interface Range
 {
