+++ tests/Driver/NgFileUploadHandlerTest.php
@@ -1,23 +1,22 @@
 <?php
 
-namespace CodingSocks\ChunkUploader\Tests\Driver;
+namespace CodingSocks\UploadHandler\Tests\Driver;
 
-use CodingSocks\ChunkUploader\Driver\SimpleUploaderJsUploadDriver;
-use CodingSocks\ChunkUploader\Event\FileUploaded;
-use CodingSocks\ChunkUploader\Exception\InternalServerErrorHttpException;
-use CodingSocks\ChunkUploader\Tests\TestCase;
-use CodingSocks\ChunkUploader\UploadHandler;
+use CodingSocks\UploadHandler\Driver\NgFileBaseHandler;
+use CodingSocks\UploadHandler\Event\FileUploaded;
+use CodingSocks\UploadHandler\Exception\InternalServerErrorHttpException;
+use CodingSocks\UploadHandler\Tests\TestCase;
+use CodingSocks\UploadHandler\UploadHandler;
 use Illuminate\Http\Request;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Event;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Validation\ValidationException;
 use Mockery;
-use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\HttpKernel\Exception\MethodNotAllowedHttpException;
 
-class SimpleUploaderUploadDriverTest extends TestCase
+class NgFileUploadHandlerTest extends TestCase
 {
     /**
      * @var UploadHandler
@@ -28,9 +27,9 @@ class SimpleUploaderUploadDriverTest extends TestCase
     {
         parent::setUp();
 
-        $this->app->make('config')->set('chunk-uploader.identifier', 'nop');
-        $this->app->make('config')->set('chunk-uploader.uploader', 'simple-uploader-js');
-        $this->app->make('config')->set('chunk-uploader.sweep', false);
+        $this->app->make('config')->set('upload-handler.identifier', 'nop');
+        $this->app->make('config')->set('upload-handler.handler', 'ng-file-upload');
+        $this->app->make('config')->set('upload-handler.sweep', false);
         $this->handler = $this->app->make(UploadHandler::class);
 
         Storage::fake('local');
@@ -39,9 +38,9 @@ class SimpleUploaderUploadDriverTest extends TestCase
 
     public function testDriverInstance()
     {
-        $manager = $this->app->make('chunk-uploader.upload-manager');
+        $manager = $this->app->make('upload-handler.upload-manager');
 
-        $this->assertInstanceOf(SimpleUploaderJsUploadDriver::class, $manager->driver());
+        $this->assertInstanceOf(NgFileBaseHandler::class, $manager->driver());
     }
 
     public function notAllowedRequestMethods()
@@ -72,40 +71,28 @@ class SimpleUploaderUploadDriverTest extends TestCase
 
     public function testResumeWhenChunkDoesNotExists()
     {
-        $this->createFakeLocalFile('chunks/200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt', '000-099');
-
         $request = Request::create('', Request::METHOD_GET, [
-            'chunkNumber' => 2,
-            'totalChunks' => 2,
-            'chunkSize' => 100,
-            'totalSize' => 200,
-            'identifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
-            'filename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'relativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'currentChunkSize' => 100,
-        ]);
-
-        $response = $this->createTestResponse($this->handler->handle($request));
-        $response->assertStatus(Response::HTTP_NO_CONTENT);
-    }
-
-    public function testResume()
-    {
-        $this->createFakeLocalFile('chunks/200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt', '000-099');
-
-        $request = Request::create('', Request::METHOD_GET, [
-            'chunkNumber' => 1,
-            'totalChunks' => 2,
-            'chunkSize' => 100,
-            'totalSize' => 200,
-            'identifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
-            'filename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'relativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'currentChunkSize' => 100,
+            'file' => 'test.txt',
+            'totalSize' => '200',
         ]);
 
         $response = $this->createTestResponse($this->handler->handle($request));
         $response->assertSuccessful();
+        $response->assertJson(['size' => 0]);
+    }
+
+    public function testResume()
+    {
+        $this->createFakeLocalFile('chunks/200_test.txt', '000-099');
+
+        $request = Request::create('', Request::METHOD_GET, [
+            'file' => 'test.txt',
+            'totalSize' => '200',
+        ]);
+
+        $response = $this->createTestResponse($this->handler->handle($request));
+        $response->assertSuccessful();
+        $response->assertJson(['size' => 100]);
     }
 
     public function testUploadWhenFileParameterIsEmpty()
@@ -133,17 +120,51 @@ class SimpleUploaderUploadDriverTest extends TestCase
         $this->handler->handle($request);
     }
 
+    public function testUploadMonolith()
+    {
+        $file = UploadedFile::fake()->create('test.txt', 100);
+        $request = Request::create('', Request::METHOD_POST, [], [], [
+            'file' => $file,
+        ]);
+
+        $response = $this->createTestResponse($this->handler->handle($request));
+        $response->assertSuccessful();
+        $response->assertJson(['done' => 100]);
+
+        Storage::disk('local')->assertExists($file->hashName('merged'));
+
+        Event::assertDispatched(FileUploaded::class, function ($event) use ($file) {
+            return $event->file = $file->hashName('merged');
+        });
+    }
+
+    public function testUploadMonolithWithCallback()
+    {
+        $file = UploadedFile::fake()->create('test.txt', 100);
+        $request = Request::create('', Request::METHOD_POST, [], [], [
+            'file' => $file,
+        ]);
+
+        $callback = $this->createClosureMock(
+            $this->once(),
+            'local',
+            $file->hashName('merged')
+        );
+
+        $this->handler->handle($request, $callback);
+
+        Event::assertDispatched(FileUploaded::class, function ($event) use ($file) {
+            return $event->file = $file->hashName('merged');
+        });
+    }
+
     public function excludedPostParameterProvider()
     {
         return [
-            'chunkNumber' => ['chunkNumber'],
-            'totalChunks' => ['totalChunks'],
-            'chunkSize' => ['chunkSize'],
-            'totalSize' => ['totalSize'],
-            'identifier' => ['identifier'],
-            'filename' => ['filename'],
-            'relativePath' => ['relativePath'],
-            'currentChunkSize' => ['currentChunkSize'],
+            '_chunkNumber' => ['_chunkNumber'],
+            '_chunkSize' => ['_chunkSize'],
+            '_totalSize' => ['_totalSize'],
+            '_currentChunkSize' => ['_currentChunkSize'],
         ];
     }
 
@@ -153,14 +174,10 @@ class SimpleUploaderUploadDriverTest extends TestCase
     public function testPostParameterValidation($exclude)
     {
         $arr = [
-            'chunkNumber' => 1,
-            'totalChunks' => 2,
-            'chunkSize' => 100,
-            'totalSize' => 200,
-            'identifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
-            'filename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'relativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'currentChunkSize' => 100,
+            '_chunkNumber' => 1,
+            '_chunkSize' => 100,
+            '_totalSize' => 200,
+            '_currentChunkSize' => 100,
         ];
 
         unset($arr[$exclude]);
@@ -179,14 +196,10 @@ class SimpleUploaderUploadDriverTest extends TestCase
     {
         $file = UploadedFile::fake()->create('test.txt', 100);
         $request = Request::create('', Request::METHOD_POST, [
-            'chunkNumber' => 1,
-            'totalChunks' => 2,
-            'chunkSize' => 100,
-            'totalSize' => 200,
-            'identifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
-            'filename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'relativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'currentChunkSize' => 100,
+            '_chunkNumber' => 0,
+            '_chunkSize' => 100,
+            '_totalSize' => 200,
+            '_currentChunkSize' => 100,
         ], [], [
             'file' => $file,
         ]);
@@ -195,7 +208,7 @@ class SimpleUploaderUploadDriverTest extends TestCase
         $response->assertSuccessful();
         $response->assertJson(['done' => 50]);
 
-        Storage::disk('local')->assertExists('chunks/200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt/000-099');
+        Storage::disk('local')->assertExists('chunks/200_test.txt/000-099');
 
         Event::assertNotDispatched(FileUploaded::class, function ($event) use ($file) {
             return $event->file = $file->hashName('merged');
@@ -206,14 +219,10 @@ class SimpleUploaderUploadDriverTest extends TestCase
     {
         $file = UploadedFile::fake()->create('test.txt', 100);
         $request = Request::create('', Request::METHOD_POST, [
-            'chunkNumber' => 1,
-            'totalChunks' => 2,
-            'chunkSize' => 100,
-            'totalSize' => 200,
-            'identifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
-            'filename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'relativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'currentChunkSize' => 100,
+            '_chunkNumber' => 0,
+            '_chunkSize' => 100,
+            '_totalSize' => 200,
+            '_currentChunkSize' => 100,
         ], [], [
             'file' => $file,
         ]);
@@ -229,18 +238,14 @@ class SimpleUploaderUploadDriverTest extends TestCase
 
     public function testUploadLastChunk()
     {
-        $this->createFakeLocalFile('chunks/200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt', '000-099');
+        $this->createFakeLocalFile('chunks/200_test.txt', '000-099');
 
         $file = UploadedFile::fake()->create('test.txt', 100);
         $request = Request::create('', Request::METHOD_POST, [
-            'chunkNumber' => 2,
-            'totalChunks' => 2,
-            'chunkSize' => 100,
-            'totalSize' => 200,
-            'identifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
-            'filename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'relativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'currentChunkSize' => 100,
+            '_chunkNumber' => 1,
+            '_chunkSize' => 100,
+            '_totalSize' => 200,
+            '_currentChunkSize' => 100,
         ], [], [
             'file' => $file,
         ]);
@@ -249,7 +254,7 @@ class SimpleUploaderUploadDriverTest extends TestCase
         $response->assertSuccessful();
         $response->assertJson(['done' => 100]);
 
-        Storage::disk('local')->assertExists('chunks/200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt/100-199');
+        Storage::disk('local')->assertExists('chunks/200_test.txt/100-199');
         Storage::disk('local')->assertExists($file->hashName('merged'));
 
         Event::assertDispatched(FileUploaded::class, function ($event) use ($file) {
@@ -259,18 +264,14 @@ class SimpleUploaderUploadDriverTest extends TestCase
 
     public function testUploadLastChunkWithCallback()
     {
-        $this->createFakeLocalFile('chunks/200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt', '000-099');
+        $this->createFakeLocalFile('chunks/200_test.txt', '000-099');
 
         $file = UploadedFile::fake()->create('test.txt', 100);
         $request = Request::create('', Request::METHOD_POST, [
-            'chunkNumber' => 2,
-            'totalChunks' => 2,
-            'chunkSize' => 100,
-            'totalSize' => 200,
-            'identifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
-            'filename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'relativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'currentChunkSize' => 100,
+            '_chunkNumber' => 1,
+            '_chunkSize' => 100,
+            '_totalSize' => 200,
+            '_currentChunkSize' => 100,
         ], [], [
             'file' => $file,
         ]);
