+++ tests/Driver/ResumableJsUploadHandlerTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace CodingSocks\ChunkUploader\Tests\Driver;
+namespace CodingSocks\UploadHandler\Tests\Driver;
 
-use CodingSocks\ChunkUploader\Driver\FlowJsUploadDriver;
-use CodingSocks\ChunkUploader\Event\FileUploaded;
-use CodingSocks\ChunkUploader\Exception\InternalServerErrorHttpException;
-use CodingSocks\ChunkUploader\Tests\TestCase;
-use CodingSocks\ChunkUploader\UploadHandler;
+use CodingSocks\UploadHandler\Driver\ResumableJsBaseHandler;
+use CodingSocks\UploadHandler\Event\FileUploaded;
+use CodingSocks\UploadHandler\Exception\InternalServerErrorHttpException;
+use CodingSocks\UploadHandler\Tests\TestCase;
+use CodingSocks\UploadHandler\UploadHandler;
 use Illuminate\Http\Request;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Event;
@@ -17,7 +17,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\HttpKernel\Exception\MethodNotAllowedHttpException;
 
-class FlowJsUploadDriverTest extends TestCase
+class ResumableJsUploadHandlerTest extends TestCase
 {
     /**
      * @var UploadHandler
@@ -28,9 +28,9 @@ class FlowJsUploadDriverTest extends TestCase
     {
         parent::setUp();
 
-        $this->app->make('config')->set('chunk-uploader.identifier', 'nop');
-        $this->app->make('config')->set('chunk-uploader.uploader', 'flow-js');
-        $this->app->make('config')->set('chunk-uploader.sweep', false);
+        $this->app->make('config')->set('upload-handler.identifier', 'nop');
+        $this->app->make('config')->set('upload-handler.handler', 'resumable-js');
+        $this->app->make('config')->set('upload-handler.sweep', false);
         $this->handler = $this->app->make(UploadHandler::class);
 
         Storage::fake('local');
@@ -39,9 +39,9 @@ class FlowJsUploadDriverTest extends TestCase
 
     public function testDriverInstance()
     {
-        $manager = $this->app->make('chunk-uploader.upload-manager');
+        $manager = $this->app->make('upload-handler.upload-manager');
 
-        $this->assertInstanceOf(FlowJsUploadDriver::class, $manager->driver());
+        $this->assertInstanceOf(ResumableJsBaseHandler::class, $manager->driver());
     }
 
     public function notAllowedRequestMethods()
@@ -75,14 +75,15 @@ class FlowJsUploadDriverTest extends TestCase
         $this->createFakeLocalFile('chunks/200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt', '000-099');
 
         $request = Request::create('', Request::METHOD_GET, [
-            'flowChunkNumber' => 2,
-            'flowTotalChunks' => 2,
-            'flowChunkSize' => 100,
-            'flowTotalSize' => 200,
-            'flowIdentifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
-            'flowFilename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'flowRelativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'flowCurrentChunkSize' => 100,
+            'resumableChunkNumber' => 2,
+            'resumableTotalChunks' => 2,
+            'resumableChunkSize' => 100,
+            'resumableTotalSize' => 200,
+            'resumableIdentifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
+            'resumableFilename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'resumableRelativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'resumableCurrentChunkSize' => 100,
+            'resumableType' => 'text/plain',
         ]);
 
         $response = $this->createTestResponse($this->handler->handle($request));
@@ -94,14 +95,15 @@ class FlowJsUploadDriverTest extends TestCase
         $this->createFakeLocalFile('chunks/200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt', '000-099');
 
         $request = Request::create('', Request::METHOD_GET, [
-            'flowChunkNumber' => 1,
-            'flowTotalChunks' => 2,
-            'flowChunkSize' => 100,
-            'flowTotalSize' => 200,
-            'flowIdentifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
-            'flowFilename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'flowRelativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'flowCurrentChunkSize' => 100,
+            'resumableChunkNumber' => 1,
+            'resumableTotalChunks' => 2,
+            'resumableChunkSize' => 100,
+            'resumableTotalSize' => 200,
+            'resumableIdentifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
+            'resumableFilename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'resumableRelativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'resumableCurrentChunkSize' => 100,
+            'resumableType' => 'text/plain',
         ]);
 
         $response = $this->createTestResponse($this->handler->handle($request));
@@ -136,14 +138,15 @@ class FlowJsUploadDriverTest extends TestCase
     public function excludedPostParameterProvider()
     {
         return [
-            'flowChunkNumber' => ['flowChunkNumber'],
-            'flowTotalChunks' => ['flowTotalChunks'],
-            'flowChunkSize' => ['flowChunkSize'],
-            'flowTotalSize' => ['flowTotalSize'],
-            'flowIdentifier' => ['flowIdentifier'],
-            'flowFilename' => ['flowFilename'],
-            'flowRelativePath' => ['flowRelativePath'],
-            'flowCurrentChunkSize' => ['flowCurrentChunkSize'],
+            'resumableChunkNumber' => ['resumableChunkNumber'],
+            'resumableTotalChunks' => ['resumableTotalChunks'],
+            'resumableChunkSize' => ['resumableChunkSize'],
+            'resumableTotalSize' => ['resumableTotalSize'],
+            'resumableIdentifier' => ['resumableIdentifier'],
+            'resumableFilename' => ['resumableFilename'],
+            'resumableRelativePath' => ['resumableRelativePath'],
+            'resumableCurrentChunkSize' => ['resumableCurrentChunkSize'],
+            'resumableType' => ['resumableType'],
         ];
     }
 
@@ -153,14 +156,15 @@ class FlowJsUploadDriverTest extends TestCase
     public function testPostParameterValidation($exclude)
     {
         $arr = [
-            'flowChunkNumber' => 1,
-            'flowTotalChunks' => 2,
-            'flowChunkSize' => 100,
-            'flowTotalSize' => 200,
-            'flowIdentifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
-            'flowFilename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'flowRelativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'flowCurrentChunkSize' => 100,
+            'resumableChunkNumber' => 1,
+            'resumableTotalChunks' => 2,
+            'resumableChunkSize' => 100,
+            'resumableTotalSize' => 200,
+            'resumableIdentifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
+            'resumableFilename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'resumableRelativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'resumableCurrentChunkSize' => 100,
+            'resumableType' => 'text/plain',
         ];
 
         unset($arr[$exclude]);
@@ -179,14 +183,15 @@ class FlowJsUploadDriverTest extends TestCase
     {
         $file = UploadedFile::fake()->create('test.txt', 100);
         $request = Request::create('', Request::METHOD_POST, [
-            'flowChunkNumber' => 1,
-            'flowTotalChunks' => 2,
-            'flowChunkSize' => 100,
-            'flowTotalSize' => 200,
-            'flowIdentifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
-            'flowFilename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'flowRelativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'flowCurrentChunkSize' => 100,
+            'resumableChunkNumber' => 1,
+            'resumableTotalChunks' => 2,
+            'resumableChunkSize' => 100,
+            'resumableTotalSize' => 200,
+            'resumableIdentifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
+            'resumableFilename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'resumableRelativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'resumableCurrentChunkSize' => 100,
+            'resumableType' => 'text/plain',
         ], [], [
             'file' => $file,
         ]);
@@ -206,14 +211,15 @@ class FlowJsUploadDriverTest extends TestCase
     {
         $file = UploadedFile::fake()->create('test.txt', 100);
         $request = Request::create('', Request::METHOD_POST, [
-            'flowChunkNumber' => 1,
-            'flowTotalChunks' => 2,
-            'flowChunkSize' => 100,
-            'flowTotalSize' => 200,
-            'flowIdentifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
-            'flowFilename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'flowRelativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'flowCurrentChunkSize' => 100,
+            'resumableChunkNumber' => 1,
+            'resumableTotalChunks' => 2,
+            'resumableChunkSize' => 100,
+            'resumableTotalSize' => 200,
+            'resumableIdentifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
+            'resumableFilename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'resumableRelativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'resumableCurrentChunkSize' => 100,
+            'resumableType' => 'text/plain',
         ], [], [
             'file' => $file,
         ]);
@@ -233,14 +239,15 @@ class FlowJsUploadDriverTest extends TestCase
 
         $file = UploadedFile::fake()->create('test.txt', 100);
         $request = Request::create('', Request::METHOD_POST, [
-            'flowChunkNumber' => 2,
-            'flowTotalChunks' => 2,
-            'flowChunkSize' => 100,
-            'flowTotalSize' => 200,
-            'flowIdentifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
-            'flowFilename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'flowRelativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'flowCurrentChunkSize' => 100,
+            'resumableChunkNumber' => 2,
+            'resumableTotalChunks' => 2,
+            'resumableChunkSize' => 100,
+            'resumableTotalSize' => 200,
+            'resumableIdentifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
+            'resumableFilename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'resumableRelativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'resumableCurrentChunkSize' => 100,
+            'resumableType' => 'text/plain',
         ], [], [
             'file' => $file,
         ]);
@@ -263,14 +270,15 @@ class FlowJsUploadDriverTest extends TestCase
 
         $file = UploadedFile::fake()->create('test.txt', 100);
         $request = Request::create('', Request::METHOD_POST, [
-            'flowChunkNumber' => 2,
-            'flowTotalChunks' => 2,
-            'flowChunkSize' => 100,
-            'flowTotalSize' => 200,
-            'flowIdentifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
-            'flowFilename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'flowRelativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
-            'flowCurrentChunkSize' => 100,
+            'resumableChunkNumber' => 2,
+            'resumableTotalChunks' => 2,
+            'resumableChunkSize' => 100,
+            'resumableTotalSize' => 200,
+            'resumableIdentifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
+            'resumableFilename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'resumableRelativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'resumableCurrentChunkSize' => 100,
+            'resumableType' => 'text/plain',
         ], [], [
             'file' => $file,
         ]);
