+++ tests/Driver/SimpleUploaderUploadHandlerTest.php
@@ -1,22 +1,23 @@
 <?php
 
-namespace CodingSocks\ChunkUploader\Tests\Driver;
+namespace CodingSocks\UploadHandler\Tests\Driver;
 
-use CodingSocks\ChunkUploader\Driver\NgFileUploadDriver;
-use CodingSocks\ChunkUploader\Event\FileUploaded;
-use CodingSocks\ChunkUploader\Exception\InternalServerErrorHttpException;
-use CodingSocks\ChunkUploader\Tests\TestCase;
-use CodingSocks\ChunkUploader\UploadHandler;
+use CodingSocks\UploadHandler\Driver\SimpleUploaderJsHandler;
+use CodingSocks\UploadHandler\Event\FileUploaded;
+use CodingSocks\UploadHandler\Exception\InternalServerErrorHttpException;
+use CodingSocks\UploadHandler\Tests\TestCase;
+use CodingSocks\UploadHandler\UploadHandler;
 use Illuminate\Http\Request;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Event;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Validation\ValidationException;
 use Mockery;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\HttpKernel\Exception\MethodNotAllowedHttpException;
 
-class NgFileUploadDriverTest extends TestCase
+class SimpleUploaderUploadHandlerTest extends TestCase
 {
     /**
      * @var UploadHandler
@@ -27,9 +28,9 @@ class NgFileUploadDriverTest extends TestCase
     {
         parent::setUp();
 
-        $this->app->make('config')->set('chunk-uploader.identifier', 'nop');
-        $this->app->make('config')->set('chunk-uploader.uploader', 'ng-file-upload');
-        $this->app->make('config')->set('chunk-uploader.sweep', false);
+        $this->app->make('config')->set('upload-handler.identifier', 'nop');
+        $this->app->make('config')->set('upload-handler.handler', 'simple-uploader-js');
+        $this->app->make('config')->set('upload-handler.sweep', false);
         $this->handler = $this->app->make(UploadHandler::class);
 
         Storage::fake('local');
@@ -38,9 +39,9 @@ class NgFileUploadDriverTest extends TestCase
 
     public function testDriverInstance()
     {
-        $manager = $this->app->make('chunk-uploader.upload-manager');
+        $manager = $this->app->make('upload-handler.upload-manager');
 
-        $this->assertInstanceOf(NgFileUploadDriver::class, $manager->driver());
+        $this->assertInstanceOf(SimpleUploaderJsHandler::class, $manager->driver());
     }
 
     public function notAllowedRequestMethods()
@@ -71,28 +72,40 @@ class NgFileUploadDriverTest extends TestCase
 
     public function testResumeWhenChunkDoesNotExists()
     {
+        $this->createFakeLocalFile('chunks/200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt', '000-099');
+
         $request = Request::create('', Request::METHOD_GET, [
-            'file' => 'test.txt',
-            'totalSize' => '200',
+            'chunkNumber' => 2,
+            'totalChunks' => 2,
+            'chunkSize' => 100,
+            'totalSize' => 200,
+            'identifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
+            'filename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'relativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'currentChunkSize' => 100,
         ]);
 
         $response = $this->createTestResponse($this->handler->handle($request));
-        $response->assertSuccessful();
-        $response->assertJson(['size' => 0]);
+        $response->assertStatus(Response::HTTP_NO_CONTENT);
     }
 
     public function testResume()
     {
-        $this->createFakeLocalFile('chunks/200_test.txt', '000-099');
+        $this->createFakeLocalFile('chunks/200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt', '000-099');
 
         $request = Request::create('', Request::METHOD_GET, [
-            'file' => 'test.txt',
-            'totalSize' => '200',
+            'chunkNumber' => 1,
+            'totalChunks' => 2,
+            'chunkSize' => 100,
+            'totalSize' => 200,
+            'identifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
+            'filename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'relativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'currentChunkSize' => 100,
         ]);
 
         $response = $this->createTestResponse($this->handler->handle($request));
         $response->assertSuccessful();
-        $response->assertJson(['size' => 100]);
     }
 
     public function testUploadWhenFileParameterIsEmpty()
@@ -120,51 +133,17 @@ class NgFileUploadDriverTest extends TestCase
         $this->handler->handle($request);
     }
 
-    public function testUploadMonolith()
-    {
-        $file = UploadedFile::fake()->create('test.txt', 100);
-        $request = Request::create('', Request::METHOD_POST, [], [], [
-            'file' => $file,
-        ]);
-
-        $response = $this->createTestResponse($this->handler->handle($request));
-        $response->assertSuccessful();
-        $response->assertJson(['done' => 100]);
-
-        Storage::disk('local')->assertExists($file->hashName('merged'));
-
-        Event::assertDispatched(FileUploaded::class, function ($event) use ($file) {
-            return $event->file = $file->hashName('merged');
-        });
-    }
-
-    public function testUploadMonolithWithCallback()
-    {
-        $file = UploadedFile::fake()->create('test.txt', 100);
-        $request = Request::create('', Request::METHOD_POST, [], [], [
-            'file' => $file,
-        ]);
-
-        $callback = $this->createClosureMock(
-            $this->once(),
-            'local',
-            $file->hashName('merged')
-        );
-
-        $this->handler->handle($request, $callback);
-
-        Event::assertDispatched(FileUploaded::class, function ($event) use ($file) {
-            return $event->file = $file->hashName('merged');
-        });
-    }
-
     public function excludedPostParameterProvider()
     {
         return [
-            '_chunkNumber' => ['_chunkNumber'],
-            '_chunkSize' => ['_chunkSize'],
-            '_totalSize' => ['_totalSize'],
-            '_currentChunkSize' => ['_currentChunkSize'],
+            'chunkNumber' => ['chunkNumber'],
+            'totalChunks' => ['totalChunks'],
+            'chunkSize' => ['chunkSize'],
+            'totalSize' => ['totalSize'],
+            'identifier' => ['identifier'],
+            'filename' => ['filename'],
+            'relativePath' => ['relativePath'],
+            'currentChunkSize' => ['currentChunkSize'],
         ];
     }
 
@@ -174,10 +153,14 @@ class NgFileUploadDriverTest extends TestCase
     public function testPostParameterValidation($exclude)
     {
         $arr = [
-            '_chunkNumber' => 1,
-            '_chunkSize' => 100,
-            '_totalSize' => 200,
-            '_currentChunkSize' => 100,
+            'chunkNumber' => 1,
+            'totalChunks' => 2,
+            'chunkSize' => 100,
+            'totalSize' => 200,
+            'identifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
+            'filename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'relativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'currentChunkSize' => 100,
         ];
 
         unset($arr[$exclude]);
@@ -196,10 +179,14 @@ class NgFileUploadDriverTest extends TestCase
     {
         $file = UploadedFile::fake()->create('test.txt', 100);
         $request = Request::create('', Request::METHOD_POST, [
-            '_chunkNumber' => 0,
-            '_chunkSize' => 100,
-            '_totalSize' => 200,
-            '_currentChunkSize' => 100,
+            'chunkNumber' => 1,
+            'totalChunks' => 2,
+            'chunkSize' => 100,
+            'totalSize' => 200,
+            'identifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
+            'filename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'relativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'currentChunkSize' => 100,
         ], [], [
             'file' => $file,
         ]);
@@ -208,7 +195,7 @@ class NgFileUploadDriverTest extends TestCase
         $response->assertSuccessful();
         $response->assertJson(['done' => 50]);
 
-        Storage::disk('local')->assertExists('chunks/200_test.txt/000-099');
+        Storage::disk('local')->assertExists('chunks/200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt/000-099');
 
         Event::assertNotDispatched(FileUploaded::class, function ($event) use ($file) {
             return $event->file = $file->hashName('merged');
@@ -219,10 +206,14 @@ class NgFileUploadDriverTest extends TestCase
     {
         $file = UploadedFile::fake()->create('test.txt', 100);
         $request = Request::create('', Request::METHOD_POST, [
-            '_chunkNumber' => 0,
-            '_chunkSize' => 100,
-            '_totalSize' => 200,
-            '_currentChunkSize' => 100,
+            'chunkNumber' => 1,
+            'totalChunks' => 2,
+            'chunkSize' => 100,
+            'totalSize' => 200,
+            'identifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
+            'filename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'relativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'currentChunkSize' => 100,
         ], [], [
             'file' => $file,
         ]);
@@ -238,14 +229,18 @@ class NgFileUploadDriverTest extends TestCase
 
     public function testUploadLastChunk()
     {
-        $this->createFakeLocalFile('chunks/200_test.txt', '000-099');
+        $this->createFakeLocalFile('chunks/200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt', '000-099');
 
         $file = UploadedFile::fake()->create('test.txt', 100);
         $request = Request::create('', Request::METHOD_POST, [
-            '_chunkNumber' => 1,
-            '_chunkSize' => 100,
-            '_totalSize' => 200,
-            '_currentChunkSize' => 100,
+            'chunkNumber' => 2,
+            'totalChunks' => 2,
+            'chunkSize' => 100,
+            'totalSize' => 200,
+            'identifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
+            'filename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'relativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'currentChunkSize' => 100,
         ], [], [
             'file' => $file,
         ]);
@@ -254,7 +249,7 @@ class NgFileUploadDriverTest extends TestCase
         $response->assertSuccessful();
         $response->assertJson(['done' => 100]);
 
-        Storage::disk('local')->assertExists('chunks/200_test.txt/100-199');
+        Storage::disk('local')->assertExists('chunks/200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt/100-199');
         Storage::disk('local')->assertExists($file->hashName('merged'));
 
         Event::assertDispatched(FileUploaded::class, function ($event) use ($file) {
@@ -264,14 +259,18 @@ class NgFileUploadDriverTest extends TestCase
 
     public function testUploadLastChunkWithCallback()
     {
-        $this->createFakeLocalFile('chunks/200_test.txt', '000-099');
+        $this->createFakeLocalFile('chunks/200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt', '000-099');
 
         $file = UploadedFile::fake()->create('test.txt', 100);
         $request = Request::create('', Request::METHOD_POST, [
-            '_chunkNumber' => 1,
-            '_chunkSize' => 100,
-            '_totalSize' => 200,
-            '_currentChunkSize' => 100,
+            'chunkNumber' => 2,
+            'totalChunks' => 2,
+            'chunkSize' => 100,
+            'totalSize' => 200,
+            'identifier' => '200-0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zftxt',
+            'filename' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'relativePath' => '0jWZTB1ZDfRQU6VTcXy0mJnL9xKMeEz3HoSPU0Zf.txt',
+            'currentChunkSize' => 100,
         ], [], [
             'file' => $file,
         ]);
